Refactor unified object shard trait and add invocation methods
- Renamed `UnifiedObjectShard` trait to `ObjectShard` for clarity.
- Added `invoke_fn` and `invoke_obj` methods to the `ObjectShard` trait for handling function and object method invocations.
- Updated type aliases for boxed and Arc-wrapped object shards to reflect the new trait name.
- Modified tests to accommodate changes in shard trait and invocation methods.
- Adjusted shard count handling in tests to use `u32` instead of `usize`.
- Removed unnecessary tracing initialization in event integration tests.
- Updated Docker Compose configuration to remove unused volume.
- Added `test-log` dependency for better test logging.
- Enhanced connection argument handling in CLI to support context loading from a specified path.

+++ data-plane/oprc-odgm/tests/collection_management_test.rs
@@ -66,7 +66,7 @@ async fn test_create_multiple_collections() {
         ("collection3", 1, 1),
     ];
 
-    let mut total_expected_shards = 0usize;
+    let mut total_expected_shards = 0 as u32;
 
     for (name, partitions, replicas) in &collections_to_create {
         let collection_req = CreateCollectionRequest {
@@ -85,7 +85,7 @@ async fn test_create_multiple_collections() {
             .await;
         assert!(result.is_ok(), "Failed to create collection {}", name);
 
-        total_expected_shards += *partitions as usize;
+        total_expected_shards += *partitions as u32;
     }
 
     // Wait for all shards to be created
@@ -169,7 +169,7 @@ async fn test_partition_counts() {
     let odgm = env.start_odgm().await.expect("Failed to start ODGM");
 
     let partition_counts = vec![1, 2, 4, 8];
-    let mut total_expected_shards = 0usize;
+    let mut total_expected_shards = 0 as u32;
 
     for partition_count in &partition_counts {
         let collection_req = CreateCollectionRequest {
@@ -192,7 +192,7 @@ async fn test_partition_counts() {
             partition_count
         );
 
-        total_expected_shards += *partition_count as usize;
+        total_expected_shards += *partition_count as u32;
     }
 
     // Wait for shards to be created
